Add help feature
Users have no way of viewing a list of available commands
within the chatbot.

Users should be able to access command usage easily without
external documentation.

Add a `help` command that returns a formatted list of all
available commands.

The help command is intuitive to use and is on the main input screen
that users will interact with.